A hashing application with using C
LRU Cache is a data structure that stores N recently used data in a linked list so that the last used one can be accessed first for quick access. In this project, the identity number, name, surname, date of birth and address information of the people will be stored on the cache. In order to determine whether a person's information is on the cache, instead of searching on the linked list, it will be placed in the hash table according to the ID information. The ID-number and the rank they are in the linked list will be kept in the hash table, thus speeding up the search process.